Added search bar and cleaned up Homepage code
